fix(webgui): ensure input fields take full width
Previously, the input fields in the Environment tab were not expanding to the full width of their parent card. This was due to a combination of issues, including a missing  class on a parent container and a  class that was constraining the width of the entire tab.

This commit resolves the issue by:

- Removing the  class from the main container in the Environment tab.
- Adding the  class to the  in the  method in the  class.
- Adding the  class to the  that contains the input fields.
- Using the  class on the input fields to make them expand to fill the available space.